The Solr Reference branch. A Solr fork geared towards scalability and performance.

+++ solr/core/src/java/org/apache/solr/schema/LongPointField.java
@@ -50,8 +50,8 @@ public class LongPointField extends PointField implements LongValueFieldType {
     try {
       if (val instanceof CharSequence) return Long.parseLong(val.toString());
     } catch (NumberFormatException e) {
-      Double v = Double.parseDouble(val.toString());
-      return v.longValue();
+      double v = Double.parseDouble(val.toString());
+      return v;
     }
     return super.toNativeType(val);
   }
