+++ solr/core/src/java/org/apache/solr/search/facet/UniqueBlockAgg.java
@@ -67,7 +67,7 @@ public abstract class UniqueBlockAgg extends UniqueAgg {
 
     @Override
     public long getNonShardValue(int slot) {
-      return (long)counts[slot];
+      return counts[slot];
     }
   }
