+++ solr/core/src/java/org/apache/solr/util/hll/BigEndianAscendingWordDeserializer.java
@@ -65,7 +65,7 @@ class BigEndianAscendingWordDeserializer implements IWordDeserializer {
         this.bytePadding = bytePadding;
 
         final int dataBytes = (bytes.length - bytePadding);
-        final long dataBits = (dataBytes * BITS_PER_BYTE);
+        final long dataBits = ((long) dataBytes * BITS_PER_BYTE);
 
         this.wordCount = (int)(dataBits/wordLength);
 
@@ -128,7 +128,7 @@ class BigEndianAscendingWordDeserializer implements IWordDeserializer {
         // First byte
         final int bitsRemainingInFirstByte = (BITS_PER_BYTE - firstByteSkipBits);
         final int bitsToConsumeInFirstByte = Math.min(bitsRemainingInFirstByte, wordLength);
-        long firstByte = (long)bytes[firstByteIndex];
+        long firstByte = bytes[firstByteIndex];
 
         // Mask off the bits to skip in the first byte.
         final long firstByteMask = ((1L << bitsRemainingInFirstByte) - 1L);
